Add files via upload
Create, test, and validate an XHTML document that has two frames in 
one column. The top frame, which must be 20 percent of the column, 
must have at least four links to other documents; the bottom frame will 
display those documents. The links must be names of cars; the 
documents must be at least five-line descriptions of the cars.