Add files via upload
Create and test an XHTML document that displays a table of football scores from 
a collegiate football conference in which the team names have one of the primary 
colors of their respective schools. The winning scores must appear larger and in 
a different font than the losing scores. The team names must be in a script font.