Add files via upload
Create and test an XHTML document that has at least a half page of text and 
that has a small box of text embedded on the left margin, with the main text 
flowing around the small box. The embedded text must appear in a smaller font 
and also must be set in italic.