Add files via upload
Create and test an XHTML document that has six short paragraphs of text that 
describe various aspects of the state in which you live. You must define three 
different paragraph styles, p1, p2, and p3. The p1 style must use left and right 
margins of 20 pixels, a background color of pink, and a foreground color of blue. 
The p2 style must use left and right margins of 30 pixels, a background color of black, and a foreground color of yellow. The p3 style must use a text indent of 1 
centimeter, a background color of green, and a foreground color of white. The 
first and fourth paragraph must use p1, the second and fifth must use p2, and 
the third and sixth must use p3.